Update instruction to create a new project
Using `--ignore-platform-reqs` is not wise, since it will cause issues
when setting up the site on an older PHP version (e.g. 7.1). However,
that was originally added, so create-project can run without `gd`
installed.

This is a proper fix, with a composer image that contains gd and the
oldest supported PHP version.